Adds
* HTTPServer.ServeStatic, seperating functionality out of HTTPServer.Router.Handlers.ServeStatic
* static/3
* Unit tests
